Fixes inaccurate zero checking multiplication bug
When both operands are power of two (e.g. 1.0 and 32.0) then
the zero signal was being set since only the mantissa of the
normalized product (without the hidden bit) was being used to
geneate the zero signal. This commit fixes that issue by checking
for zero on the product signal which includes the hidden bits